+++ packages/types/package.json
@@ -4,7 +4,6 @@
   "main": "index.ts",
   "types": "index.ts",
   "dependencies": {
-    "@evervault/card-validator": "workspace:^",
     "jss": "^10.10.0",
     "typescript": "latest"
   }
